Added .gitignore, basic layout complete
Using GIMP I found that the left logo section is about 37.5% of the width so I set it to a flex of 3, and the right section to a flex of 5.  Found that on the right side, the upper and lower sections are each about 35%, while the center section is about 30% of the height.  So I set the upper and lower sections to flex of 7 and the middle section to a flex of 6.  Compared with design image and it's close enough to move forward!